Continues work in unifying socket implementation
