+++ src/socket/mod.rs
@@ -1,3 +1,3 @@
+pub mod socket_manager_async;
 pub mod socket_manager;
-mod async_socket_manager;
 pub mod ethernet_packet_wrapper;
