+++ src/socket/socket_manager.rs
@@ -1,47 +1,53 @@
-use std::sync::Arc;
+use pnet::datalink::{self, Channel, DataLinkReceiver, DataLinkSender, NetworkInterface};
+use tokio::sync::watch::{self, Sender};
 
-use pnet::{
-    datalink::{self, Channel, DataLinkReceiver, DataLinkSender, NetworkInterface},
-    packet::ethernet::{EthernetPacket, MutableEthernetPacket},
-};
-use tokio::sync::{watch, Mutex};
+use super::ethernet_packet_wrapper::EthernetPacketWrapper;
 
-use super::{async_socket_manager::AsyncSocketManagerImpl, ethernet_packet_wrapper::EthernetPacketWrapper};
-pub trait SocketManager {
-    fn acquire_read(&self) -> watch::Receiver<EthernetPacketWrapper>;
-}
-
-pub struct SocketManagerImpl {
-    _channel_rx: watch::Receiver<EthernetPacketWrapper>,
-    _impl: Arc<Mutex<AsyncSocketManagerImpl>>
+pub(super) struct SocketManagerImpl {
+    ethernet_tx: Box<dyn DataLinkSender>,
+    ethernet_rx: Box<dyn DataLinkReceiver>,
+    tx: Sender<EthernetPacketWrapper>,
 }
 
 impl SocketManagerImpl {
-    pub fn new(network_interface: &NetworkInterface) -> Self {
-        let empty: Vec<u8> = vec![];
-        let (channel_tx, channel_rx) = watch::channel(EthernetPacketWrapper::new(&empty));
-        let _impl = AsyncSocketManagerImpl::new(network_interface, channel_tx);
-        let socket_manager: SocketManagerImpl = SocketManagerImpl {
-            _channel_rx: channel_rx,
-            _impl: Arc::from(Mutex::new(_impl))
+    fn setup_socket(
+        network_interface: &NetworkInterface,
+    ) -> (Box<dyn DataLinkSender>, Box<dyn DataLinkReceiver>) {
+        let channel = match datalink::channel(&network_interface, Default::default()) {
+            Ok(Channel::Ethernet(tx, rx)) => (tx, rx),
+            Ok(_) => panic!("Unhandled channel type"),
+            Err(e) => panic!(
+                "An error occurred when creating the datalink channel: {}",
+                e
+            ),
+        };
+
+        return channel;
+    }
+
+    pub fn new(network_interface: &NetworkInterface, tx: Sender<EthernetPacketWrapper>) -> Self {
+        let (ethernet_tx, ethernet_rx) = Self::setup_socket(network_interface);
+        let socket_manager = SocketManagerImpl {
+            ethernet_tx: ethernet_tx,
+            ethernet_rx: ethernet_rx,
+            tx: tx,
         };
 
         return socket_manager;
     }
-}
 
-impl SocketManager for SocketManagerImpl {
-    fn acquire_read(&self) -> watch::Receiver<EthernetPacketWrapper> {
-        return self._channel_rx.clone();
-    }
-}
-
-impl SocketManagerImpl {
     pub fn start(&mut self) {
-      let clone = self._impl.clone();
-      tokio::spawn(async move {
-        let mut acquired = clone.lock().await;
-        acquired.start();
-      });
+        loop {
+            match self.ethernet_rx.next() {
+                Ok(packet) => {
+                    println!("Got packet of size: {}", packet.len());
+                    let _ = self.tx.send(EthernetPacketWrapper::new(packet));
+                }
+                Err(e) => {
+                    // If an error occurs, we can handle it here
+                    panic!("An error occurred while reading: {}", e);
+                }
+            }
+        }
     }
 }
